# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peexoo's API swagger documentation",
        "description": "This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "adds a new Booking",
                "operationId": "addBookings",
                "requestBody": {
                    "description": "Booking Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Booking"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/bookings/album": {
            "put": {
                "tags": [
                    "bookings"
                ],
                "summary": "adds a new album to booking",
                "operationId": "addAlbumToBooking",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
        "/api/bookings/report/{who}": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "Allows a user to report a photographer for a booking",
                "operationId": "reportPhotographer",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
        "/api/bookings/{booking_id}/accept": {
            "patch": {
                "tags": [
                    "bookings"
                ],
                "summary": "Accepts a booking",
                "operationId": "acceptBooking",
                "parameters": [
                    {
                        "name": "booking_id",
                        "in": "path",
                        "description": "Id of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/bookings/{booking_id}/reject": {
            "patch": {
                "tags": [
                    "bookings"
                ],
                "summary": "Rejects a booking",
                "operationId": "rejectBooking",
                "parameters": [
                    {
                        "name": "booking_id",
                        "in": "path",
                        "description": "Id of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/bookings/images/search": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "Searches for images from bookings",
                "operationId": "searchImages",
                "requestBody": {
                    "description": "Booking images object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhotographerSearchParams"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/faq": {
            "get": {
                "tags": [
                    "faq"
                ],
                "summary": "fetch all FAQ",
                "operationId": "getFAQs",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            },
            "post": {
                "tags": [
                    "faq"
                ],
                "summary": "adds a new FAQ",
                "operationId": "addFAQ",
                "requestBody": {
                    "description": "FAQ Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FAQ"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/faq/{id}": {
            "patch": {
                "tags": [
                    "faq"
                ],
                "summary": "update a new FAQ",
                "operationId": "updateFAQ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Faq Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "FAQ Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FAQ"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/photographer/report": {
            "post": {
                "tags": [
                    "photographer"
                ],
                "summary": "Report a Photographers",
                "operationId": "reportPhotographer",
                "requestBody": {
                    "description": "Photographer Oject",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Photographer"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {
                "title": "Photographer Object Representation",
                "description": "Photographer model",
                "required": [
                    "user_id",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "The id of the booking",
                        "description": "Booking Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "The Id of the user for this booking",
                        "description": "User's Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "photographer_id": {
                        "title": "Photographer's Id",
                        "description": "The Id of the photographer for this booking.",
                        "type": "integer"
                    },
                    "booking_category_id": {
                        "title": "Booking Category Id",
                        "description": "The Id of the category that the booking belongs to",
                        "type": "integer"
                    },
                    "title": {
                        "title": "Title",
                        "description": "The title for this booking. This is optional",
                        "type": "string"
                    },
                    "extra_message": {
                        "title": "Extra message",
                        "description": "Additional Text provided by the user",
                        "type": "string"
                    },
                    "country": {
                        "title": "Country",
                        "description": "The country for the booking's shoot",
                        "type": "string"
                    },
                    "state": {
                        "title": "State",
                        "description": "The state in the country for the booking's shoot",
                        "type": "string"
                    },
                    "address1": {
                        "title": "Address 1",
                        "description": "The User's first Address",
                        "type": "string"
                    },
                    "address2": {
                        "title": "Address 2",
                        "description": "The User's second Address ????????",
                        "type": "string"
                    },
                    "type": {
                        "title": "Type",
                        "description": "The Type of Booking. This is optional",
                        "type": "string"
                    },
                    "shoot_start_date": {
                        "title": "Start Date",
                        "description": "The start date of the shoot for this booking",
                        "type": "string"
                    },
                    "shoot_end_date": {
                        "title": "End Date",
                        "description": "The end date of the shoot for this booking",
                        "type": "string"
                    },
                    "delivery_date": {
                        "title": "Delivery date",
                        "description": "The Date for delivery of the booking's shoot. Defaults to shoot end date if not set.",
                        "type": "string"
                    },
                    "status": {
                        "title": "Status",
                        "description": "The current status os the booking. This is set by the backend",
                        "type": "string"
                    },
                    "package_name": {
                        "title": "Package Name",
                        "description": "The name of the package the booking belongs to",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Photographer"
                }
            },
            "FAQ": {
                "title": "FAQ Object Representation",
                "description": "FAQ model",
                "properties": {
                    "question": {
                        "title": "Question",
                        "description": "Question",
                        "type": "string"
                    },
                    "answer": {
                        "title": "Answer",
                        "description": "Answer",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "FAQ"
                }
            },
...
            "Photographer": {
                "title": "Photographer Object Representation",
                "description": "Photographer model",
                "required": [
                    "user_id",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Photographer Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "User Id",
                        "description": "User's Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "about_us": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "category": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "photography_type": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "region": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "business_name": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "verified": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "archived": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "bvn": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "bvn_verified": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "id_card": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    },
                    "id_card_verified": {
                        "title": "Introduction Text",
                        "description": "A Brief Description.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Photographer"
                }
            },
...
            "PhotographerSearchParams": {
                "title": "PhotographerSearchParams Object Representation",
                "description": "PhotographerSearchParams model",
                "required": [
                    "user_id",
                    "password"
                ],
                "properties": {
                    "categories": {
                        "title": "categories",
                        "description": "Comma seperated values of booking types",
                        "type": "string"
                    },
                    "type": {
                        "title": "type",
                        "description": "Comma seperated Photographer types",
                        "type": "string"
                    },
                    "region": {
                        "title": "region",
                        "description": "Comma seperated values of regions",
                        "type": "string"
                    },
                    "availability": {
                        "title": "availability",
                        "description": "Availability date range",
                        "type": "string"
                    },
                    "price": {
                        "title": "price",
                        "description": "Photographer price range",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PhotographerSearchParams"
                }
            },
...
        },
        "securitySchemes": {
            "api_key_scheme": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}